chore: merge multidut pfc folder
