--- tests/snappi_tests/multidut/pfc/test_multidut_global_pause_with_snappi.py
@@ -1,96 +1,0 @@
-import pytest
-import logging
-from tests.common.helpers.assertions import pytest_require, pytest_assert                            # noqa: F401
-from tests.common.fixtures.conn_graph_facts import conn_graph_facts, fanout_graph_facts, \
-    fanout_graph_facts_multidut     # noqa: F401
-from tests.common.snappi_tests.snappi_fixtures import snappi_api_serv_ip, snappi_api_serv_port, \
-    snappi_api, get_snappi_ports, is_snappi_multidut, \
-    get_snappi_ports_single_dut, snappi_testbed_config, \
-    get_snappi_ports_multi_dut, snappi_dut_base_config, cleanup_config, get_snappi_ports_for_rdma  # noqa: F401
-from tests.common.snappi_tests.qos_fixtures import lossless_prio_list, prio_dscp_map, disable_pfcwd  # noqa: F401
-from tests.snappi_tests.multidut.pfc.files.multidut_helper import run_pfc_test                      # noqa: F401
-from tests.common.snappi_tests.snappi_test_params import SnappiTestParams
-from tests.snappi_tests.variables import MULTIDUT_PORT_INFO, MULTIDUT_TESTBED
-
-logger = logging.getLogger(__name__)
-
-pytestmark = [pytest.mark.topology('multidut-tgen', 'tgen')]
-
-
-@pytest.mark.parametrize("multidut_port_info", MULTIDUT_PORT_INFO[MULTIDUT_TESTBED])
-def test_global_pause(snappi_api,                                   # noqa: F811
-                      conn_graph_facts,                             # noqa: F811
-                      fanout_graph_facts_multidut,                # noqa: F811
-                      get_snappi_ports,                           # noqa: F811
-                      duthosts,
-                      prio_dscp_map,                                # noqa: F811
-                      lossless_prio_list,                           # noqa: F811
-                      tbinfo,                                      # noqa: F811
-                      multidut_port_info,
-                      disable_pfcwd):                              # noqa: F811
-    """
-    Test if IEEE 802.3X pause (a.k.a., global pause) will impact any priority
-
-    Args:
-        snappi_api (pytest fixture): SNAPPI session
-        conn_graph_facts (pytest fixture): connection graph
-        fanout_graph_facts_multidut (pytest fixture): fanout graph for multiple duts
-        get_snappi_ports (pytest fixture): list of snappi port and duthost information
-        duthosts (pytest fixture): list of DUTs
-        lossless_prio_list (pytest fixture): list of all the lossless priorities
-        prio_dscp_map (pytest fixture): priority vs. DSCP map (key = priority).
-        tbinfo (pytest fixture): fixture provides information about testbed
-        get_snappi_ports (pytest fixture): gets snappi ports and connected DUT port info and returns as a list
-    Returns:
-        N/A
-    """
-    snappi_port_list = get_snappi_ports
-
-    tbname = tbinfo.get('conf-name', None)
-
-    tx_port_count = 1
-    rx_port_count = 1
-
-    pytest_assert(len(snappi_port_list) >= tx_port_count + rx_port_count,
-                  "Need Minimum of 2 ports defined in ansible/files/*links.csv file")
-
-    for testbed_subtype, rdma_ports in multidut_port_info.items():
-        pytest_assert(len(rdma_ports['tx_ports']) >= tx_port_count,
-                      'MULTIDUT_PORT_INFO doesn\'t have the required Tx ports defined for \
-                      testbed {}, subtype {} in variables.py'.
-                      format(MULTIDUT_TESTBED, testbed_subtype))
-
-        pytest_assert(len(rdma_ports['rx_ports']) >= rx_port_count,
-                      'MULTIDUT_PORT_INFO doesn\'t have the required Rx ports defined for \
-                      testbed {}, subtype {} in variables.py'.
-                      format(tbname, testbed_subtype))
-        logger.info('Running test for testbed subtype: {}'.format(testbed_subtype))
-        if is_snappi_multidut(duthosts):
-            snappi_ports = get_snappi_ports_for_rdma(snappi_port_list, rdma_ports,
-                                                     tx_port_count, rx_port_count, MULTIDUT_TESTBED)
-        else:
-            snappi_ports = snappi_port_list
-        testbed_config, port_config_list, snappi_ports = snappi_dut_base_config(duthosts,
-                                                                                snappi_ports,
-                                                                                snappi_api)
-
-    all_prio_list = prio_dscp_map.keys()
-    test_prio_list = lossless_prio_list
-    bg_prio_list = [x for x in all_prio_list if x not in test_prio_list]
-
-    snappi_extra_params = SnappiTestParams()
-    snappi_extra_params.multi_dut_params.multi_dut_ports = snappi_ports
-    run_pfc_test(api=snappi_api,
-                 testbed_config=testbed_config,
-                 port_config_list=port_config_list,
-                 conn_data=conn_graph_facts,
-                 fanout_data=fanout_graph_facts_multidut,
-                 global_pause=True,
-                 pause_prio_list=None,
-                 test_prio_list=test_prio_list,
-                 bg_prio_list=bg_prio_list,
-                 prio_dscp_map=prio_dscp_map,
-                 test_traffic_pause=False,
-                 snappi_extra_params=snappi_extra_params)
-
-    cleanup_config(duthosts, snappi_ports)
