+++ tests/snappi_tests/pfc/test_lossless_response_to_external_pause_storms.py
@@ -9,7 +9,7 @@ from tests.common.snappi_tests.snappi_fixtures import snappi_api_serv_ip, snappi
 from tests.common.snappi_tests.qos_fixtures import prio_dscp_map, \
     lossless_prio_list, disable_pfcwd                                                       # noqa: F401
 from tests.snappi_tests.files.helper import multidut_port_info, setup_ports_and_dut, reboot_duts  # noqa: F401
-from tests.snappi_tests.multidut.pfc.files.lossless_response_to_external_pause_storms_helper import (
+from tests.snappi_tests.pfc.files.lossless_response_to_external_pause_storms_helper import (
      run_lossless_response_to_external_pause_storms_test,
     )
 from tests.common.snappi_tests.snappi_test_params import SnappiTestParams
