+++ tests/snappi_tests/pfc/test_pfc_pause_response_with_snappi.py
@@ -8,7 +8,7 @@ from tests.common.fixtures.conn_graph_facts import conn_graph_facts,\
 from tests.common.snappi_tests.snappi_fixtures import snappi_api_serv_ip, snappi_api_serv_port,\
     snappi_api, snappi_testbed_config       # noqa F401
 from tests.common.snappi_tests.qos_fixtures import prio_dscp_map, all_prio_list, lossless_prio_list,\
-    lossy_prio_list                         # noqa F401
+    lossy_prio_list, disable_pfcwd          # noqa F401
 from tests.common.snappi_tests.snappi_test_params import SnappiTestParams
 
 logger = logging.getLogger(__name__)
@@ -26,7 +26,8 @@ def test_pfc_single_lossless_headroom(snappi_api,                       # noqa F
                                       enum_dut_lossless_prio,
                                       all_prio_list,                    # noqa F811
                                       prio_dscp_map,                    # noqa F811
-                                      enum_pfc_pause_delay_test_params):
+                                      enum_pfc_pause_delay_test_params,
+                                      disable_pfcwd):  # noqa F811
     """
     Test headroom capacity for DUT for a single lossless priority
 
@@ -99,7 +100,8 @@ def test_pfc_pause_multi_lossless_headroom(snappi_api,                  # noqa F
                                            lossless_prio_list,          # noqa F811
                                            lossy_prio_list,             # noqa F811
                                            prio_dscp_map,               # noqa F811
-                                           enum_pfc_pause_delay_test_params):
+                                           enum_pfc_pause_delay_test_params,
+                                           disable_pfcwd):   # noqa F811
     """
     Test headroom capacity for DUT for multiple lossless priorities
