+++ tests/snappi_tests/pfc/test_valid_pfc_frame_with_snappi.py
@@ -8,7 +8,7 @@ from tests.common.fixtures.conn_graph_facts import conn_graph_facts,\
 from tests.common.snappi_tests.snappi_fixtures import snappi_api_serv_ip, snappi_api_serv_port,\
     snappi_api, snappi_testbed_config # noqa F401
 from tests.common.snappi_tests.qos_fixtures import prio_dscp_map, all_prio_list, lossless_prio_list,\
-    lossy_prio_list # noqa F401
+    lossy_prio_list, disable_pfcwd # noqa F401
 from tests.common.snappi_tests.snappi_test_params import SnappiTestParams
 from tests.common.snappi_tests.common_helpers import packet_capture
 
@@ -25,7 +25,8 @@ def test_valid_pfc_frame(snappi_api, # noqa F811
                          rand_one_dut_portname_oper_up,
                          lossless_prio_list, # noqa F811
                          lossy_prio_list, # noqa F811
-                         prio_dscp_map): # noqa F811
+                         prio_dscp_map, # noqa F811
+                         disable_pfcwd): # noqa F811
     """
     Test if PFC Pause frame generated by device under test (DUT) is valid.
